move docgen Header class to a separate module
